bugs and supplydemandclass
fixed bugs in crewReplacer_Job. added an new thing that might not be vary usefull, but its not enuth for its own mod yet, and i need to put it somewere were others can use it because when it is useful it makes life easyer. bla bla bla words.

+++ jars/src/data/scripts/crewReplacer_Crew.java
@@ -80,7 +80,7 @@ public class crewReplacer_Crew {
         if(numberOfItems > 1){
             //list_commodity
             displayName = displayName + "s";
-            text.appendToLastParagraph(numberOfItems + " " + displayName);
+            text.appendToLastParagraph((int)numberOfItems + " " + displayName);
             //text.highlightInLastPara(highlight, displayName);
         }else if(numberOfItems == 1){
             //displayName = name;
